+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- MediaCenter to the image upload on `ComponentEditor`
 ### Changed
 - Accordion behavior on the array fields.
+- Animation on accordion items.
 
 ## [1.10.0] - 2018-6-28
 ### Changed